Move EncodedVariableInterpreter, Grep, and LogSurgeonReader sources into the clp namespace.
