+++ components/core/src/clp/LogSurgeonReader.hpp
@@ -1,10 +1,11 @@
-#ifndef LOG_SURGEON_READER_HPP
-#define LOG_SURGEON_READER_HPP
+#ifndef CLP_LOG_SURGEON_READER_HPP
+#define CLP_LOG_SURGEON_READER_HPP
 
 #include <log_surgeon/Reader.hpp>
 
-#include "ReaderInterface.hpp"
+#include "../ReaderInterface.hpp"
 
+namespace clp {
 /*
  * Wrapper providing a read function that works with the parsers in log_surgeon.
  */
@@ -15,5 +16,6 @@ public:
 private:
     ReaderInterface& m_reader_interface;
 };
+}  // namespace clp
 
-#endif  // LOG_SURGEON_READER_HPP
+#endif  // CLP_LOG_SURGEON_READER_HPP
