+++ components/core/src/clp/streaming_archive/writer/File.cpp
@@ -1,6 +1,6 @@
 #include "File.hpp"
 
-#include "../../../EncodedVariableInterpreter.hpp"
+#include "../../EncodedVariableInterpreter.hpp"
 
 using std::string;
 using std::to_string;
